move useFilters into own file
