+++ src/hooks/index.js
@@ -1,5 +1,6 @@
 export { useAsyncStorage } from './useAsyncStorage';
 export { useAuthToken } from './useAuthToken';
+export { useFilters } from './useFilters';
 export { useSnackbar } from './useSnackbar';
 export { useWordlistEntriesCreate } from './useWordlistEntriesCreate';
 export { useWordlistEntryId } from './useWordlistEntryId';
